fixed silly mistakes

+++ app/models.py
@@ -16,9 +16,9 @@ class User(UserMixin, db.Model):
     id = db.Column(db.Integer, primary_key=True)
     username = db.Column(db.String(64), index=True, unique=True)
     email = db.Column(db.String(120), index=True, unique=True)
-    cashBalance = db.Column(db.Integer, index = True)
-    bankBalance = db.Column(db.Integer, index = True)
-    payappBalance = db.Column(db.Integer, index = True)
+    cashBalance = db.Column(db.Float, index = True)
+    bankBalance = db.Column(db.Float, index = True)
+    payappBalance = db.Column(db.Float, index = True)
     password_hash = db.Column(db.String(128))
 
     #creating relations between user and transactions
@@ -41,7 +41,7 @@ class CashTransaction(db.Model):
     user_id = db.Column(db.Integer, db.ForeignKey('user.id')) #The user that the transaction is associated with.
     date = db.Column(db.String, index = True)
     debit = db.Column(db.Boolean, index = True) # True if the amount is to be subtracted, false if it is actually a credit
-    amount = db.Column(db.Integer, index = True)
+    amount = db.Column(db.Float, index = True)
     description = db.Column(db.String(255), index = True)
 
 class BankTransaction(db.Model):
@@ -49,7 +49,7 @@ class BankTransaction(db.Model):
     user_id = db.Column(db.Integer, db.ForeignKey('user.id')) #The user that the transaction is associated with.
     date = db.Column(db.String, index = True)
     debit = db.Column(db.Boolean, index = True) # True if the amount is to be subtracted, false if it is actually a credit
-    amount = db.Column(db.Integer, index = True)
+    amount = db.Column(db.Float, index = True)
     description = db.Column(db.String(255), index = True)
 
 class PayAppTransaction(db.Model):
@@ -57,5 +57,5 @@ class PayAppTransaction(db.Model):
     user_id = db.Column(db.Integer, db.ForeignKey('user.id')) #The user that the transaction is associated with.
     date = db.Column(db.String, index = True)
     debit = db.Column(db.Boolean, index = True) # True if the amount is to be subtracted, false if it is actually a credit
-    amount = db.Column(db.Integer, index = True)
+    amount = db.Column(db.Float, index = True)
     description = db.Column(db.String(255), index = True)
